Fix: error en repositories :bug:

+++ Canchitas/src/main/java/org/ncapas/canchitas/repositories/EstadoDisponibilidadRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EstadoDisponibilidadRepository extends JpaRepository<EstadoDisponibilidad, Long> {
+public interface EstadoDisponibilidadRepository extends JpaRepository<EstadoDisponibilidad, Integer> {
 }
