+++ Canchitas/src/main/java/org/ncapas/canchitas/repositories/SemanaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SemanaRepository extends JpaRepository<Semana, Long> {
+public interface SemanaRepository extends JpaRepository<Semana, Integer> {
 }
