+++ Canchitas/src/main/java/org/ncapas/canchitas/repositories/TipoCanchaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoCanchaRepository extends JpaRepository<TipoCancha, Long> {
+public interface TipoCanchaRepository extends JpaRepository<TipoCancha, Integer> {
 }
